fixed the date issue
i fixed the date issue by grouping the columns by date and position. However, now we need to make a for loop that assess whether each date is in the dataframe in our date range bc off days wont have data and we need to account for that. If the date is not there, we need to insert 3 rows w F M and B data. When creating the new row of data, you need the date and whether it was a W_L at the following game. Otherwise, every observation can be 0.